+++ src/lib/B9/Content/Builtin.hs
@@ -4,29 +4,29 @@ module B9.Content.Builtin where
 
 import Control.Monad.Trans (lift)
 import Control.Parallel.Strategies
-import Data.Binary
-import Data.Data
-import Data.Hashable
-import GHC.Generics (Generic)
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative
 #endif
+import Control.Monad.IO.Class
+import Data.Binary
+import qualified Data.ByteString as Strict
+import qualified Data.ByteString.Base64 as B64
+import qualified Data.ByteString.Lazy.Char8 as Lazy
+import Data.Data
+import Data.Hashable
+import GHC.Generics (Generic)
+import System.Exit
+import System.Process
+import Test.QuickCheck
+
 import B9.B9Monad
 import B9.Content.AST
 import B9.Content.CloudConfigYaml
 import B9.Content.ErlangPropList
+import B9.Content.Generator
 import B9.Content.StringTemplate
 import B9.Content.YamlObject
-import Control.Monad.IO.Class
-import qualified Data.ByteString as BS
-import qualified Data.ByteString.Base64 as B64
-import qualified Data.ByteString.Char8 as B
-import System.Exit
-
 import B9.QCUtil
-import Test.QuickCheck
-
-import System.Process
 
 data Content
   = RenderErlang (AST Content ErlangPropList)
@@ -37,7 +37,7 @@ data Content
     -- | The data in the given file will be base64 encoded.
   | RenderBase64BinaryFile FilePath
     -- | This data will be base64 encoded.
-  | RenderBase64Binary B.ByteString
+  | RenderBase64Binary Lazy.ByteString
   | FromURL String
   deriving (Read, Show, Typeable, Eq, Data, Generic)
 
@@ -56,7 +56,7 @@ instance Arbitrary Content where
       , RenderYamlObject <$> smaller arbitrary
       , RenderCloudConfig <$> smaller arbitrary
       , FromString <$> smaller arbitrary
-      , RenderBase64Binary . BS.pack <$> smaller arbitrary
+      , RenderBase64Binary . Lazy.pack <$> smaller arbitrary
       , FromURL <$> smaller arbitrary
       ]
 
@@ -68,10 +68,10 @@ instance CanRender Content where
   render (RenderBase64BinaryFile s) = readBinaryFileAsBase64 s
       -- | Read a binary file and encode it as base64
     where
-      readBinaryFileAsBase64 :: MonadIO m => FilePath -> m B.ByteString
-      readBinaryFileAsBase64 f = B64.encode <$> liftIO (B.readFile f)
-  render (RenderBase64Binary b) = return (B64.encode b)
-  render (FromString str) = return (B.pack str)
+      readBinaryFileAsBase64 :: MonadIO m => FilePath -> m Lazy.ByteString
+      readBinaryFileAsBase64 f = Lazy.fromStrict . B64.encode <$> liftIO (Strict.readFile f)
+  render (RenderBase64Binary b) = return (Lazy.fromStrict $ B64.encode $ Lazy.toStrict b)
+  render (FromString str) = return (Lazy.pack str)
   render (FromURL url) =
     lift $ do
       dbgL $ "Downloading: " ++ url
@@ -80,7 +80,7 @@ instance CanRender Content where
         then do
           dbgL $ "Download finished. Bytes read: " ++ show (length out)
           traceL $ "Downloaded (truncated to first 4K): \n\n" ++ take 4096 out ++ "\n\n"
-          return $ B.pack out
+          return $ Lazy.pack out
         else do
           errorL $ "Download failed: " ++ err
           liftIO $ exitWith exitCode