+++ src/lib/B9/Content/YamlObject.hs
@@ -8,14 +8,16 @@ import           Control.Applicative
 import           Control.Parallel.Strategies
 import           Data.Binary                 (Binary (..))
 import qualified Data.Binary.Get             as Binary
-import qualified Data.ByteString.Lazy.Char8  as B
+import qualified Data.ByteString.Lazy.Char8  as Lazy
 import           Data.Data
 import           Data.Function
 import           Data.Hashable
 import           Data.HashMap.Strict         hiding (singleton)
 import           Data.Semigroup
-import qualified Data.Text                   as T
-import qualified Data.Text.Encoding          as E
+import qualified Data.Text                   as StrictT
+import qualified Data.Text.Encoding          as StrictE
+import qualified Data.Text.Lazy              as LazyT
+import qualified Data.Text.Lazy.Encoding     as LazyE
 import           Data.Vector                 as Vector (singleton, (++))
 import           Data.Yaml                   as Yaml
 import           GHC.Generics                (Generic)
@@ -23,6 +25,7 @@ import           Prelude                     hiding ((++))
 import           Text.Printf
 
 import           B9.Content.AST
+import           B9.Content.Generator
 import           B9.Content.StringTemplate
 
 import           Test.QuickCheck
@@ -35,7 +38,7 @@ newtype YamlObject = YamlObject
 
 instance Binary YamlObject where
   put = put . encode . _fromYamlObject
-  get = YamlObject . either (error . show) id . Yaml.decodeThrow . B.toStrict <$> Binary.getRemainingLazyByteString
+  get = YamlObject . either (error . show) id . Yaml.decodeThrow . Lazy.toStrict <$> Binary.getRemainingLazyByteString
 
 instance Read YamlObject where
   readsPrec _ = readsYamlObject
@@ -44,10 +47,10 @@ instance Read YamlObject where
       readsYamlObject s = [(yamlFromString y, r2) | ("YamlObject", r1) <- lex s, (y, r2) <- reads r1]
         where
           yamlFromString :: String -> YamlObject
-          yamlFromString = either error id . decodeSyntax "HERE-DOC" . E.encodeUtf8 . T.pack
+          yamlFromString = either error id . decodeSyntax "HERE-DOC" . LazyE.encodeUtf8 . LazyT.pack
 
 instance Show YamlObject where
-  show (YamlObject o) = "YamlObject " <> show (T.unpack $ E.decodeUtf8 $ encode o)
+  show (YamlObject o) = "YamlObject " <> show (StrictT.unpack $ StrictE.decodeUtf8 $ encode o)
 
 instance Semigroup YamlObject where
   (YamlObject v1) <> (YamlObject v2) = YamlObject (combine v1 v2)
@@ -74,7 +77,7 @@ instance ASTish YamlObject where
       ASTMerge asts -> do
         ys <- mapM fromAST asts
         return (foldl1 (<>) ys)
-      ASTEmbed c -> YamlObject . toJSON . T.unpack . E.decodeUtf8 <$> render c
+      ASTEmbed c -> YamlObject . toJSON . StrictT.unpack . StrictE.decodeUtf8 . Lazy.toStrict <$> render c
       ASTString str -> return (YamlObject (toJSON str))
       ASTInt int -> return (YamlObject (toJSON int))
       ASTParse src@(Source _ srcPath) -> do
@@ -86,7 +89,7 @@ instance ASTish YamlObject where
     where
       fromASTPair (key, value) = do
         (YamlObject o) <- fromAST value
-        let key' = T.pack key
+        let key' = StrictT.pack key
         return $ key' .= o
 
 instance Arbitrary YamlObject where